requirements.txt
pyyaml
pyyaml>=5.4.1
requests>=2.25.1